load-uninitialized-blocks: move logic and tests from sync_manager to unfinalized_tracker
