+++ crates/consensus-logic/src/unfinalized_tracker.rs
@@ -2,6 +2,7 @@
 
 use std::collections::*;
 
+use alpen_vertex_db::traits::L2DataProvider;
 use alpen_vertex_primitives::buf::Buf32;
 use alpen_vertex_state::prelude::*;
 
@@ -189,6 +190,28 @@ impl UnfinalizedBlockTracker {
         })
     }
 
+    /// Loads the unfinalized blocks into the tracker which are already in the DB
+    pub fn load_unfinalized_blocks(
+        &mut self,
+        finalized_height: u64,
+        database: &impl L2DataProvider,
+    ) -> anyhow::Result<()> {
+        let mut height = finalized_height;
+        while let Ok(block_ids) = database.get_blocks_at_height(height) {
+            if block_ids.is_empty() {
+                break;
+            }
+            for block_id in block_ids {
+                if let Some(block) = database.get_block_data(block_id)? {
+                    let header = block.header();
+                    let _ = self.attach_block(block_id, header);
+                }
+            }
+            height += 1;
+        }
+        Ok(())
+    }
+
     #[cfg(test)]
     pub fn unchecked_set_finalized_tip(&mut self, id: L2BlockId) {
         self.finalized_tip = id;
@@ -247,3 +270,94 @@ impl FinalizeReport {
 }
 
 // TODO unit tests
+
+#[cfg(test)]
+mod tests {
+    use alpen_test_utils::ArbitraryGenerator;
+    use alpen_vertex_db::traits::{Database, L2DataStore};
+    use alpen_vertex_state::{
+        block::{L2Block, L2BlockBody, L2BlockHeader},
+        id::L2BlockId,
+    };
+
+    use crate::unfinalized_tracker;
+
+    fn get_genesis_block() -> L2Block {
+        let arb = ArbitraryGenerator::new();
+        let mut header: L2BlockHeader = arb.generate();
+        let empty_hash = L2BlockId::default();
+        let body: L2BlockBody = arb.generate();
+        header.set_parent_and_idx(empty_hash, 0);
+        L2Block::new(header, body)
+    }
+
+    fn get_mock_block_with_parent(parent: &L2BlockHeader) -> L2Block {
+        let arb = ArbitraryGenerator::new();
+        let mut header: L2BlockHeader = arb.generate();
+        let body: L2BlockBody = arb.generate();
+        header.set_parent_and_idx(parent.get_blockid(), parent.blockidx() + 1);
+        L2Block::new(header, body)
+    }
+
+    fn setup_test_chain(
+        l2_prov: &impl L2DataStore,
+    ) -> (L2BlockHeader, L2BlockHeader, L2BlockHeader, L2BlockHeader) {
+        // b2   b2a (side chain)
+        // |   /
+        // | /
+        // b1 (finalized)
+        // |
+        // g1 (10)
+        // |
+
+        let genesis = get_genesis_block();
+        let genesis_header = genesis.header().clone();
+
+        let block1 = get_mock_block_with_parent(genesis.header());
+        let block1_header = block1.header().clone();
+
+        let block2 = get_mock_block_with_parent(block1.header());
+        let block2_header = block2.header().clone();
+
+        let block2a = get_mock_block_with_parent(block1.header());
+        let block2a_header = block2a.header().clone();
+
+        l2_prov.put_block_data(genesis.clone()).unwrap();
+        l2_prov.put_block_data(block1.clone()).unwrap();
+        l2_prov.put_block_data(block2.clone()).unwrap();
+        l2_prov.put_block_data(block2a.clone()).unwrap();
+
+        (genesis_header, block1_header, block2_header, block2a_header)
+    }
+
+    #[test]
+    fn test_load_unfinalized_blocks() {
+        let db = alpen_test_utils::get_common_db();
+        let l2_prov = db.l2_provider();
+
+        let (genesis, block1, block2, block2a) = setup_test_chain(l2_prov.as_ref());
+
+        // Init the chain tracker from the state we figured out.
+        let mut chain_tracker =
+            unfinalized_tracker::UnfinalizedBlockTracker::new_empty(genesis.get_blockid());
+
+        chain_tracker
+            .load_unfinalized_blocks(1, l2_prov.as_ref())
+            .unwrap();
+
+        assert_eq!(
+            chain_tracker.get_parent(&block1.get_blockid()),
+            Some(&genesis.get_blockid())
+        );
+
+        assert_eq!(
+            chain_tracker.get_parent(&block2.get_blockid()),
+            Some(&block1.get_blockid())
+        );
+
+        assert_eq!(
+            chain_tracker.get_parent(&block2a.get_blockid()),
+            Some(&block1.get_blockid())
+        );
+    }
+}
